create varchar module
